Add flow-copy-source script to enable flowtype to users
We have to wait and see what FlowType will offer in the future. But now we go w/ that hack where each ES6 ist stored twice in the build folder - once minified and production ready as ES5 code and once with .flow extension as is. That of course bloats the packaged library. But what to do? I rather have flow type definitions. Supposedly there is an experimental script being developed that let's one extract all FlowType definitions from a project and ship that instead.