Temporarily render empty span in RawButton
Currently, if you pass a styled-component a single child, it freaks out. It tries to convert that symbol into a string. Yet if you pass more than 1 child to the component it works. I am sure this will be resolved in the near future. But we can't wait for it. We go w/ this workaround.